admin: fix "Git Repositories" link
The DefaultSimpleLinkManager had been encountering NullPointerExceptions trying to create the link in the sidebar.
When it encounters an exception, it logs it, and then ignores the link.
It looks like perhaps the section we were adding to doesn't exist any more; I changed to the section that the SVN plugin was using.